チャット画面スクロール対応追加
